LPS-191346 Simplify
